requirements.txt
pillow>=9.0.0
rich>=13.0.0 